# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/file/avatar/{username}": {
      "get": {
        "operationId": "FileController_getUserAvatar",
        "summary": "User's avatar",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "backColor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image"
          }
        },
        "tags": [
          "file"
        ]
      }
    },
...
    "/file/{filename}": {
      "get": {
        "operationId": "FileController_getAttachment",
        "summary": "Get an uploaded file by it's name",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File content"
          }
        },
        "tags": [
          "file"
        ]
      }
    },
    "/file/stats/{filename}": {
      "get": {
        "operationId": "FileController_getAttachmentInfo",
        "summary": "Get info of an uploaded file by it's name",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File stats"
          }
        },
        "tags": [
          "file"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Forgot password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "string"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/validate-reset-password-key": {
      "post": {
        "operationId": "AuthController_validateKey",
        "summary": "Validate the reset password key",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthValidateKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "true"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Reset password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "message": "true"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/chats/{conversationId}": {
      "get": {
        "operationId": "ChatsController_getFromConversation",
        "summary": "Get all messages from a conversation",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/create/{userId}": {
      "post": {
        "operationId": "ChatsController_createConversation",
        "summary": "Create a conversation with a user",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/send/{receiverId}": {
      "post": {
        "operationId": "ChatsController_sendMessgae",
        "summary": "Send message with HTTP",
        "parameters": [
          {
            "name": "receiverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "chats"
        ]
      }
    },
...
  },
  "info": {
    "title": "Attendance API",
    "description": "API documentation for attendance backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "AuthLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "MANAGEMENT",
              "MARKETING",
              "DEVELOPMENT",
              "SALES",
              "RND"
            ]
          }
        },
        "required": [
          "name",
          "username",
          "email",
          "password",
          "role"
        ]
      },
      "AuthForgotPasswordDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ]
      },
      "AuthValidateKeyDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "key"
        ]
      },
      "AuthResetPasswordDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "key",
          "password"
        ]
      },
...
      "MessageEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "conversationId": {
            "type": "number"
          },
          "content": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "createdAt",
          "conversationId",
          "content",
          "userId"
        ]
      },
      "ConversationEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "members",
          "type",
          "createdAt"
        ]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "reciever": {
            "type": "string"
          },
          "recieverId": {
            "type": "number"
          },
          "conversationId": {
            "type": "number"
          },
          "groupId": {
            "type": "number"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "reciever",
          "recieverId",
          "conversationId",
          "groupId",
          "attachments"
        ]
      },
...
    }
  }
}